Adding bookmarks.  Unforeseen difficulties.  Want them to be cross file capable.  Have to keep track of bookmark removals and document closings.  Just need to sleep on it.

+++ src/docman.d
@@ -342,7 +342,7 @@ class DOCMAN
 		if(Current !is null)return Current.Word;
 		return null;
 	}
-	ulong GetLineNo()
+	int GetLineNo()
 	{
 		if(Current !is null)return Current.LineNumber;
 		return -1;
